Which medications to avoid in people at risk of delirium: a systematic review
Background: delirium is a common clinical problem and is associated with adverse health outcomes. Many medications have been associated with the development of delirium, but the strength of the associations is uncertain and it is unclear which medications should be avoided in people at risk of delirium.
Methods: we conducted a systematic review to identify prospective studies that investigated the association between medications and risk of delirium. A sensitivity analysis was performed to construct an evidence hierarchy for the risk of delirium with individual agents.
Results: a total of 18,767 studies were identified by the search strategy. Fourteen studies met the inclusion criteria.
Delirium risk appears to be increased with opioids (odds ratio [OR] 2.5, 95% CI 1.2–5.2), benzodiazepines (3.0, 1.3–6.8), dihydropyridines (2.4, 1.0–5.8) and possibly antihistamines (1.8, 0.7–4.5). There appears to be no increased risk with neuroleptics (0.9, 0.6–1.3) or digoxin (0.5, 0.3–0.9). There is uncertainty regarding H2 antagonists, tricyclic antidepressants, antiparkinson medications, steroids, non-steroidal anti-inflammatory drugs and antimuscarinics.
Conclusion: for people at risk of delirium, avoid new prescriptions of benzodiazepines or consider reducing or stopping these medications where possible. Opioids should be prescribed with caution in people at risk of delirium, but this should be tempered by the observation that untreated severe pain can itself trigger delirium. Caution is also required when prescribing dihydropyridines and antihistamine H1 antagonists for people at risk of delirium and considered individual patient assessment is advocated.