Delirium and Long-term Cognitive Trajectory Among Persons With Dementia
Conclusions: Delirium is highly prevalent among persons with AD who are hospitalized and is associated with an increased rate of cognitive deterioration that is maintained for up to 5 years. Strategies to prevent delirium may represent a promising avenue to explore for ameliorating cognitive deterioration in AD.

Arch Intern Med. 2012;172(17):1324-1331.
Published online August 20, 2012.
doi:10.1001/archinternmed.2012.3203